Interact with Objects
You can now interact with objects. Objects now exist, in fact. In my testing scene there is a cube, you can ponder it, you can interact with it, and the character says something different with ALL of these interactions depending on the difficulty (able to be set by the difficulty handler gameobject currently).