Test for sequential objects
A cube that unlocks a "door." The player can't interact with objects if he's mumbling to himself (this is to prevent the player from missing out on interaction text if he's thinking while interacting. I'm probably going to implement a way to skip dialogue soon, so it's easier on the player).