Barebones Inventory System
Implemented an item system, where you can create items. Currently, the player is given items by interacting with an object inside the game world that has the Additem() function inside of their OnInteract() function, so when the player interacts with that object, it gives the player the item and then destroys itself. The player can use a key (capsule) to open a door (huge square) and it seems to be working fine, the item gets removed once it's used and whatnot. The item doesn't appear in the players hotbar, and it's an automatic use kind of deal. This means that the player doesn't have to equip items to use them, but instead they're used automatically if the object that needs that item finds them while going through the players inventory. Oh yeah I renamed a couple things and changed some dialogue. I also added UI elements for future use.